requirements.txt
jupyter
numpy
pandas
matplotlib
seaborn
scikit-learn
pyaml
openpyxl
us
plotly
geopandas
dash
python-dotenv